moved the generation strategies to own module
