+++ src/redisearch_rs/qint/Cargo.toml
@@ -4,10 +4,23 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 
-[dev-dependencies]
-proptest = { workspace = true, features = ["std"] }
-proptest-derive.workspace = true
+[features]
+test-extensions = ["dep:criterion", "dep:proptest", "dep:proptest-derive"]
+
+[lib]
+name = "qint"
+path = "src/lib.rs"
+
+[[test]]
+name = "qint-tests"
+path = "tests/qint.rs"
+required-features = ["test-extensions"]
+
+[dependencies]
 rand.workspace = true
+criterion = { workspace = true, optional = true}
+proptest = { workspace = true, features = ["std"], optional = true }
+proptest-derive = { workspace = true, optional = true }
 
 [lints]
 workspace = true
