+++ src/redisearch_rs/qint/src/lib.rs
@@ -250,3 +250,6 @@ where
         }
     }
 }
+
+#[cfg(feature = "test-extensions")]
+pub mod test_utils;
